search: factor out query values and delete query/types
