--- internal/search/query/types/doc.go
@@ -1,3 +1,0 @@
-// Package types contains facilities for checking the validity
-// and typesafety of search queries.
-package types
